Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
	modified:   sparsegpt.py

+++ sparsegpt.py
@@ -629,8 +629,7 @@ class SparseGPT:
                         Probmask_solve(temp_model, 0.5, train_loader, self.dev)
                     # Calculate loss
                     # temp_model.weight.data = temp_model.weight.data.to(self.layer.weight.data.dtype)
-                    temp_model.weight.data = temp_model.weight.data.to(dtype)
-                    current_loss = torch.sum((temp_model(self.inp1) - self.out1) ** 2).item()
+                    current_loss = torch.sum((temp_model(self.inp1.to(torch.float32)) - self.out1.to(torch.float32)) ** 2).item()
 
                     # Update best hyperparameters if current loss is lower
                     if current_loss < min_loss:
@@ -638,7 +637,7 @@ class SparseGPT:
                         best_lr = lr
                         best_rho = rho
                         best_max_iter = max_iter
-                        self.layer.weight.data = temp_model.subnet * temp_model.weight.data
+                        self.layer.weight.data = (temp_model.subnet * temp_model.weight.data).to(dtype)
 
         # Print the best hyperparameters and the corresponding loss
         print(f"Best lr: {best_lr}, Best rho: {best_rho}, Best max_iter: {best_max_iter}, Minimum Loss: {min_loss}")
